Move some graph search methods to a helper class
